+++ packages/venia-concept/src/components/Navigation/navHeader.css
@@ -3,5 +3,6 @@
     display: inline-flex;
     font-size: 1rem;
     font-weight: 400;
+    padding: 0 0.5rem;
     text-transform: uppercase;
 }
