+++ packages/venia-concept/src/components/Navigation/navHeader.js
@@ -1,38 +1,54 @@
-import React, { Component, Fragment } from 'react';
-import { func, shape, string } from 'prop-types';
+import React, { Fragment } from 'react';
+import { bool, func, shape, string } from 'prop-types';
+import {
+    ArrowLeft as ArrowLeftIcon,
+    Menu as MenuIcon,
+    X as CloseIcon
+} from 'react-feather';
 
-import classify from 'src/classify';
+import { mergeClasses } from 'src/classify';
 import Icon from 'src/components/Icon';
-import { ArrowLeft as ArrowLeftIcon, X as CloseIcon } from 'react-feather';
 import Trigger from 'src/components/Trigger';
 import defaultClasses from './navHeader.css';
 
-class NavHeader extends Component {
-    static propTypes = {
-        classes: shape({
-            title: string
-        }),
-        onBack: func.isRequired,
-        onClose: func.isRequired
-    };
+const titles = {
+    CREATE_ACCOUNT: 'Create Account',
+    FORGOT_PASSWORD: 'Forgot Password',
+    MY_ACCOUNT: 'My Account',
+    SIGN_IN: 'Sign In',
+    MENU: 'Main Menu'
+};
 
-    render() {
-        const { classes, onBack, onClose, title } = this.props;
+const NavHeader = props => {
+    const { isTopLevel, onBack, onClose, view } = props;
+    const classes = mergeClasses(defaultClasses, props.classes);
 
-        return (
-            <Fragment>
-                <Trigger key="backButton" action={onBack}>
-                    <Icon src={ArrowLeftIcon} />
-                </Trigger>
-                <h2 key="title" className={classes.title}>
-                    <span>{title}</span>
-                </h2>
-                <Trigger key="closeButton" action={onClose}>
-                    <Icon src={CloseIcon} />
-                </Trigger>
-            </Fragment>
-        );
-    }
-}
+    const title = titles[view] || titles.MENU;
+    const backIcon = isTopLevel && view === 'MENU' ? MenuIcon : ArrowLeftIcon;
 
-export default classify(defaultClasses)(NavHeader);
+    return (
+        <Fragment>
+            <Trigger key="backButton" action={onBack}>
+                <Icon src={backIcon} />
+            </Trigger>
+            <h2 key="title" className={classes.title}>
+                <span>{title}</span>
+            </h2>
+            <Trigger key="closeButton" action={onClose}>
+                <Icon src={CloseIcon} />
+            </Trigger>
+        </Fragment>
+    );
+};
+
+export default NavHeader;
+
+NavHeader.propTypes = {
+    classes: shape({
+        title: string
+    }),
+    isTopLevel: bool,
+    onBack: func.isRequired,
+    onClose: func.isRequired,
+    view: string.isRequired
+};
